# [Xamarin.Android.Build.Tasks] Debug-mode linker improvements