@jonathanpeppers 2019-01-04
Build logs: [Logs.zip](https://github.com/xamarin/xamarin-android/files/2728835/Logs.zip)

Mainly sending this to "see what breaks", I want to test it a little more, hence the `do-not-merge` label.

I also want to measure the impact to startup time for this.