@jonathanpeppers 2019-01-10
I'm going to rework this, and only do the `StripEmbeddedLibraries` portion for now.