@jonathanpeppers 2019-01-10
So there is a bit of slowdown here:
```
Before:
 554 ms  LinkAssemblies                             1 calls
After:
1248 ms  LinkAssemblies                             1 calls
```
If we save ~200ms on deploy time, and it looks like it saves ~10ms in startup time (from what I measured on a Pixel 2) this tradeoff isn't quite worth it.

I will close this for now, and come back to it if it makes sense.